Move `FieldTypes` to `io.spine.code.proto`
